+++ lib/shopify-cli/tasks/ensure_env.rb
@@ -21,7 +21,7 @@ module ShopifyCli
           api_key: api_key,
           secret: api_secret,
           shop: shop,
-          host: Tasks::Tunnel.call(@ctx),
+          host: ShopifyCli::Tunnel.start(@ctx),
           scopes: 'write_products,write_customers,write_draft_orders',
         ).write(@ctx)
         env
