+++ test/project_types/node/commands/serve_test.rb
@@ -14,7 +14,7 @@ module Node
       end
 
       def test_server_command
-        ShopifyCli::Tasks::Tunnel.stubs(:call)
+        ShopifyCli::Tunnel.stubs(:start)
         ShopifyCli::Tasks::UpdateDashboardURLS.expects(:call)
         ShopifyCli::Helpers::EnvFile.any_instance.expects(:update)
         @context.expects(:system).with(
@@ -33,7 +33,7 @@ module Node
 
       def test_open_while_run
         ShopifyCli::Context.any_instance.stubs(:on_siginfo).yields
-        ShopifyCli::Tasks::Tunnel.stubs(:call).returns('https://example.com')
+        ShopifyCli::Tunnel.stubs(:start).returns('https://example.com')
         ShopifyCli::Tasks::UpdateDashboardURLS.expects(:call)
         ShopifyCli::Helpers::EnvFile.any_instance.expects(:update).with(
           @context, :host, 'https://example.com'
@@ -46,7 +46,7 @@ module Node
       end
 
       def test_update_env_with_host
-        ShopifyCli::Tasks::Tunnel.expects(:call).never
+        ShopifyCli::Tunnel.expects(:start).never
         ShopifyCli::Tasks::UpdateDashboardURLS.expects(:call)
         ShopifyCli::Helpers::EnvFile.any_instance.expects(:update).with(
           @context, :host, 'https://example-foo.com'
