+++ test/project_types/rails/commands/serve_test.rb
@@ -14,7 +14,7 @@ module Rails
       end
 
       def test_server_command
-        ShopifyCli::Tasks::Tunnel.stubs(:call)
+        ShopifyCli::Tunnel.stubs(:start)
         ShopifyCli::Tasks::UpdateDashboardURLS.expects(:call)
         ShopifyCli::Helpers::EnvFile.any_instance.expects(:update)
         @context.expects(:system).with(
@@ -32,7 +32,7 @@ module Rails
 
       def test_open_while_run
         ShopifyCli::Context.any_instance.stubs(:on_siginfo).yields
-        ShopifyCli::Tasks::Tunnel.stubs(:call).returns('https://example.com')
+        ShopifyCli::Tunnel.stubs(:start).returns('https://example.com')
         ShopifyCli::Tasks::UpdateDashboardURLS.expects(:call)
         ShopifyCli::Helpers::EnvFile.any_instance.expects(:update).with(
           @context, :host, 'https://example.com'
@@ -45,7 +45,7 @@ module Rails
       end
 
       def test_update_env_with_host
-        ShopifyCli::Tasks::Tunnel.expects(:call).never
+        ShopifyCli::Tunnel.expects(:start).never
         ShopifyCli::Tasks::UpdateDashboardURLS.expects(:call)
         ShopifyCli::Helpers::EnvFile.any_instance.expects(:update).with(
           @context, :host, 'https://example-foo.com'
